+++ src/components/page-body/trip-events-component.js
@@ -1,14 +1,10 @@
-import AbstractComponent from "./abstract-component";
+import AbstractComponent from "../abstract-component";
 
 export default class TripEventsComponent extends AbstractComponent {
   getTemplate() {
     return (
       `<section class="trip-events">
         <h2 class="visually-hidden">Trip events</h2>
-
-        <!-- Сортировка -->
-
-        <!-- Контент -->
       </section>`
     );
   }
