+++ src/components/page-header/trip-info-component.js
@@ -1,4 +1,4 @@
-import AbstractComponent from './abstract-component';
+import AbstractComponent from '../abstract-component';
 
 const createTripInfoTemplate = (totalCost) => {
   return (
