+++ src/models/points-model.js
@@ -1,4 +1,4 @@
-import {FilterTypes} from '../components/filters-component';
+import {FilterTypes} from '../components/page-header/filters-component';
 import {replaceElementInArray, removeFromArray, unshiftArray} from '../utils/common';
 import FilterPredicate from '../utils/filter';
